extracted SymptomTab to its own file
