+++ app/src/main/java/com/example/theperiodpurse/ui/calendar/CalendarScreen.kt
@@ -45,6 +45,7 @@ import com.example.theperiodpurse.Screen
 import com.example.theperiodpurse.ui.theme.HeaderColor1
 import com.example.theperiodpurse.ui.theme.SelectedColor1
 import com.example.theperiodpurse.data.Symptom
+import com.example.theperiodpurse.ui.calendar.components.SymptomTab
 import com.example.theperiodpurse.ui.calendar.components.getDayColorAndIcon
 import com.example.theperiodpurse.ui.theme.ThePeriodPurseTheme
 import com.google.accompanist.pager.*
@@ -63,118 +64,6 @@ val tabModifier = Modifier
     .background(Color.White)
     .fillMaxWidth()
 
-@Composable
-fun SymptomTab(
-    trackedSymptoms: List<Symptom>,
-    selectedSymptom: Symptom,
-    onSymptomClick: (Symptom) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    var expanded by remember { mutableStateOf(false) }
-    Column(modifier = modifier) {
-        DisplaySymptomTab(
-            selectedSymptom = selectedSymptom,
-            expanded = expanded,
-            onExpandButtonClick = { expanded = !expanded },
-            modifier = tabModifier
-        )
-        if (expanded) {
-            SwitchSymptomTab(
-                selectedSymptom = selectedSymptom,
-                symptoms = trackedSymptoms,
-                onSymptomOnClicks = trackedSymptoms.map { { onSymptomClick(it) } },
-                modifier = tabModifier
-            )
-        }
-    }
-}
-
-@Composable
-private fun DisplaySymptomTab(
-    selectedSymptom: Symptom,
-    expanded: Boolean,
-    onExpandButtonClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.Center,
-        modifier = modifier
-    ) {
-        Text(
-            text = stringResource(selectedSymptom.nameId),
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier.padding(end = 2.dp)
-        )
-        Icon(
-            painter = painterResource(
-                id = selectedSymptom.resourceId
-            ),
-            tint = Color.Black,
-            contentDescription = selectedSymptom.name,
-            modifier = Modifier
-                .padding(end = 0.dp)
-                .testTag("Selected Symptom")
-        )
-        SwitchSymptomButton(
-            expanded = expanded,
-            onClick = onExpandButtonClick
-        )
-    }
-}
-
-@Composable
-private fun SwitchSymptomButton(
-    expanded: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    IconButton(onClick = onClick, modifier = modifier) {
-        Icon(
-            imageVector = if (expanded) Icons.Filled.ExpandLess else Icons.Filled.ExpandMore,
-            tint = Color.Gray,
-            contentDescription = stringResource(R.string.expand_button_symptoms_content_description)
-        )
-    }
-}
-
-@Composable
-private fun SwitchSymptomTab(
-    selectedSymptom: Symptom,
-    symptoms: List<Symptom>,
-    onSymptomOnClicks: List<() -> Unit>,
-    modifier: Modifier = Modifier
-) {
-    Row(
-        horizontalArrangement = Arrangement.SpaceEvenly,
-        modifier = modifier.testTag("Symptom Options")
-    ) {
-        symptoms.zip(onSymptomOnClicks).forEach { (symptom, onClick) ->
-            val interactionSource = remember { MutableInteractionSource() }
-            val isPressed by interactionSource.collectIsPressedAsState()
-
-            IconButton(
-                onClick = onClick,
-                interactionSource = interactionSource
-            ) {
-                val defaultColor = Color.Black
-                val color = if (isPressed) {
-                    defaultColor.copy(ContentAlpha.disabled)
-                } else if (selectedSymptom == symptom) {
-                    Color(0xFFBF3428)
-                } else {
-                    defaultColor
-                }
-                Icon(
-                    painter = painterResource(id = symptom.resourceId),
-                    tint = color,
-                    contentDescription = stringResource(symptom.nameId)
-                )
-            }
-        }
-    }
-}
 
 inline fun Modifier.noRippleClickable(crossinline onClick: () -> Unit): Modifier = composed {
     clickable(indication = null,
